Implementing liskov principle
Instead of committing the service layer w/ a specific database class, we use a generic interface, so this way it's possible to change database more easily in the future.

And also, this will stay truth w/ Liskov principle in the sense that any repositry implemented for now on should implements that structure.